requirements.txt
markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
djangorestframework==3.11.2
django-filter==2.2.0
django-rest-swagger==2.2.0
django==3.0
django-crum==0.7.5
django-import-export>=2.1.0
loducode_utils==1.5.1
psycopg2-binary==2.8.4
# setuptools
# wheel
# twine